feat: language menu from wpml active langs, separating table
