+++ admin/src/assets/styles/common/_variables.scss
@@ -49,6 +49,7 @@ $white: #fff;
 $box-shadow-small:  0 1px 3px rgba($black-real, 0.16);
 $br-10: 0.625rem;
 
+$breakpoint-desktop-hd: 1900px;
 $breakpoint-desktop-big: 1600px;
 $breakpoint-desktop-box: 1380px;
 $breakpoint-desktop: 1180px;
